Add posibility ot setup linger option of AnyEvent::Handle object
This option is in effect when, for example, code terminates connection
by calling "disconnect" but there are ongoing operations. In this case
destructor of underlying AnyEvent::Handle object will keep the write
buffer in memory for long time (1 hour by default). If the amount
of sending data is big it causes severe memory leak.